requirements.txt
jinja2
python-jenkins
pyapi-gitlab
pyyaml
requests
